# Add fast CIDR.Contains implementation