Before:
```text
$ go test -benchmem -bench=BenchmarkCIDRContains -benchtime=10s ./wgcfg
goos: windows
goarch: amd64
pkg: github.com/tailscale/wireguard-go/wgcfg
BenchmarkCIDRContainsIPv4/IPv4-8                138954278               85.2 ns/op            52 B/op          2 allocs/op
BenchmarkCIDRContainsIPv4/IPv6-8                100000000              102 ns/op              64 B/op          2 allocs/op
PASS
ok      github.com/tailscale/wireguard-go/wgcfg 31.010s
```

After:
```
$ go test -benchmem -bench=BenchmarkCIDRContains -benchtime=10s ./wgcfg
goos: windows
goarch: amd64
pkg: github.com/tailscale/wireguard-go/wgcfg
BenchmarkCIDRContainsIPv4/IPv4-8                1000000000              11.7 ns/op             0 B/op          0 allocs/op
BenchmarkCIDRContainsIPv4/IPv6-8                721590703               16.6 ns/op             0 B/op          0 allocs/op
PASS
ok      github.com/tailscale/wireguard-go/wgcfg 26.666s
```

Fixes https://github.com/tailscale/tailscale/issues/73.