@crawshaw 2020-03-10
Thanks! When I PR wgcfg upstream (soon I hope), I'll keep this commit intact.